--- Home/src/app/about/page.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-import { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: "About",
-  description: "Generated by create next app",
-};
-
-const AboutUs = () => {
-  return (
-    <div>
-      <h2>This page contains information about us</h2>
-    </div>
-  );
-};
-
-export default AboutUs;
