--- Home/src/app/cart/page.tsx
@@ -1,17 +1,0 @@
-import { Metadata } from "next";
-import { CartUsers } from "@/components/cartrafce";
-
-export const metadata: Metadata = {
-  title: "Cart",
-  description: "Generated by create next app",
-};
-
-const Cart = () => {
-  return (
-    <>
-      <CartUsers />
-    </>
-  );
-};
-
-export default Cart;
