+++ Home/src/app/layout.tsx
@@ -1,6 +1,6 @@
 import "../styles/globals.css";
 import { Footer, Nav } from "@/components/sharedLayoutsrafce";
-import { CartProvider, HomeProvider } from "@/contextrafce";
+import { HomeProvider } from "@/contextrafce";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 
@@ -8,7 +8,7 @@ const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
   title: "Home",
-  description: "Generated by create next app",
+  description: "This is the home page",
 };
 
 export default function RootLayout({
@@ -20,11 +20,9 @@ export default function RootLayout({
     <html lang="en">
       <body className={inter.className}>
         <HomeProvider>
-          <CartProvider>
-            <Nav />
-            {children}
-            <Footer />
-          </CartProvider>
+          <Nav />
+          {children}
+          <Footer />
         </HomeProvider>
       </body>
     </html>
