+++ Home/src/assets/logos/index.ts
@@ -1,0 +1,3 @@
+import storeFrontLogo from "./StoreFront.png";
+
+export { storeFrontLogo };
