--- Home/src/components/cart/index.ts
@@ -1,3 +1,0 @@
-import CartUsers from "./cartUsers";
-
-export { CartUsers };
