+++ Home/src/components/general/logo/logo.module.css
@@ -1,0 +1,4 @@
+.div {
+  /* border: 2px solid green; */
+  width: fit-content;
+}
