+++ Home/src/components/sharedLayouts/footer/footer.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../../styles/sharedLayoutStyles/footer.module.css";
+import styles from "./footer.module.css";
 
 const Footer = () => {
   return (
