+++ Home/src/components/sharedLayouts/nav/nav.module.css
@@ -1,10 +1,10 @@
 .nav {
   border: 2px solid gray;
-  margin: 20px;
-  width: fit-content;
+  max-width: 1300px;
+  margin-inline: auto;
 }
 
 .nav a {
-  margin-inline: 20px;
+  /* margin-inline: 20px; */
   font-size: var(--medium);
 }
