+++ Home/src/context/index.ts
@@ -1,6 +1,4 @@
 import { HomeProvider } from "./homeContext";
-import { CartProvider } from "./cartContext";
 import { useHomeContext } from "./homeContext";
-import { useCartContext } from "./cartContext";
 
-export { HomeProvider, CartProvider, useCartContext, useHomeContext };
+export { HomeProvider, useHomeContext };
