--- Home/src/styles/cartStyles/cartUser.module.css
@@ -1,4 +1,0 @@
-.div p {
-  font-size: var(--medium);
-  margin: 20px;
-}
