+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "axios": "^1.6.2",
     "react-icons": "^4.12.0"
   }
 }
